Reworked restoring saved passwords
Saved ids are ignored, the coords are used instead as discriminators. When the password-protected tile is interacted with, the coords will be removed from all other irrelevant objects. This is the dirty fix to bypass the inaccurate saved objIds.

The ids are inaccurate because the original code assumes a different implementation. (So that id of opened door is saved now, but most people leave closed doors instead)