Spawn to a random seed from a list instead of default spawn
When players try to spawn as eve without a seed spawn them to a random seed from defaultSeedList.ini.
The list is processed by the server and the players dont have access to it. It may be possible for secret seeds to be added to it.

Warning: comes on as default. Can be disabled by useSeedList.ini > 0.

It just works.